Move some functions into different files
